// src/types/const.ts
export interface Benefit {
  title: string;
  text: string;
}

export interface Project {
  id: string;
  image: string;
  alt: string;
  title: string;
}

export interface Testimonial {
  review: string;
  name: string;
  designation: string;
  image: string;
}

export interface FAQ {
  question: string;
  answer: string;
}